Pass to double since it can actually calculate really big factorials (e.g, 22!)

+++ DiscreteTechniques/TecnicasConteo.cs
@@ -12,7 +12,7 @@ namespace Tecnicas
             mt = new MathTools();
         }
 
-        public UInt64 permutacionSinRepeticion(UInt64 n)
+        public double permutacionSinRepeticion(double n)
         {
             return mt.factorial(n);
         }
@@ -23,11 +23,11 @@ namespace Tecnicas
         /// <param name="n">Cantidad de elementos que contiene</param>
         /// <param name="subconjuntos">Una lista que contiene subconjuntos cuya suma es n</param>
         /// <returns></returns>
-        public UInt64 permutacionConRepeticion(UInt64 n, List<UInt64> subconjuntos)
+        public double permutacionConRepeticion(double n, List<double> subconjuntos)
         {
-            UInt64 denominador = 1;
+            double denominador = 1;
 
-            foreach(UInt64 m in subconjuntos)
+            foreach(double m in subconjuntos)
             {
                 denominador *= mt.factorial(m);
                 Console.WriteLine(m);
@@ -39,46 +39,46 @@ namespace Tecnicas
             return mt.factorial(n) / denominador;
         }
 
-        public UInt64 variacionSinRepeticion(UInt64 n, UInt64 m)
+        public double variacionSinRepeticion(double n, double m)
         {
-            UInt64 divisor = mt.factorial(n - m);
+            double divisor = mt.factorial(n - m);
 
             return mt.factorial(n) / divisor;
         }
 
-        public UInt64 variacionConRepeticion(UInt64 n, UInt64 m)
+        public double variacionConRepeticion(double n, double m)
         {
-            return (UInt64)Math.Pow(n, m);
+            return (double)Math.Pow(n, m);
         }
 
-        public UInt64 combinacionSinRepeticion(UInt64 n, UInt64 m)
+        public double combinacionSinRepeticion(double n, double m)
         {
-            UInt64 divisor = mt.factorial(m) * mt.factorial(n - m);
+            double divisor = mt.factorial(m) * mt.factorial(n - m);
 
             return mt.factorial(n) / divisor;
         }
 
-        public UInt64 combinacionConRepeticion(UInt64 n, UInt64 m)
+        public double combinacionConRepeticion(double n, double m)
         {
-            UInt64 numerador = mt.factorial(m + n - 1);
-            UInt64 denominador = mt.factorial(n) * mt.factorial(m - 1);
+            double numerador = mt.factorial(m + n - 1);
+            double denominador = mt.factorial(n) * mt.factorial(m - 1);
 
             Console.WriteLine("Num: " + numerador + " Dem: " + denominador);
 
             return numerador / denominador;
         }
 
-        public UInt64 permutacionCircular(UInt64 n)
+        public double permutacionCircular(double n)
         {
             return mt.factorial(n - 1);
         }
 
-        public UInt64 reglaMultiplicativa(UInt64 k)
+        public double reglaMultiplicativa(double k)
         {
             return 0;
         }
 
-        public UInt64 reglaAditiva()
+        public double reglaAditiva()
         {
             return 0;
         }
@@ -91,7 +91,7 @@ namespace Tecnicas
 
         }
 
-        public UInt64 factorial(UInt64 numero)
+        public double factorial(double numero)
         {
             if (numero == 0)
                 return 1;
